+++ src/app/[locale]/layout.module.css
